Added Isomorphic Material Relay Starter Kit
I added it to the boilerplate section. While it is more than just a boilerplate because it contains examples, they are fairly easy to remove and it could be used as a boilerplate too. I believe this is the most appropriate section, however please move it if you feel otherwise.