union animation components to a different "component animation" category
It's much easier this way to find animation components, and look much cleaner and tidy. 
The component section is way too big and messy anyway.
And the more splitted it'll be the less conflicts, cause there will be less chance that two people will change the same line this way.