feat: Added Tolgee
Added Tolgee to the Internationalization section.

Reason: Tolgee is open-source and can be self-hosted to translate a React app

About Tolgee:
Modify texts right in the React app. In-context translation that doesn't translate lines of text separately. Machine translation is available on top of a lot of tools for teamwork